Move dynamic-configuration into its own crate
