+++ Cargo.toml
@@ -12,6 +12,7 @@ members = [
   "ddcommon-ffi",
   "ddtelemetry",
   "ddtelemetry-ffi",
+  "dynamic-configuration",
   "tools",
   "ipc",
   "ipc/macros",
