+++ dynamic-configuration/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+edition = "2021"
+license = "Apache 2.0"
+name = "datadog-dynamic-configuration"
+version = "0.0.1"
+
+[features]
+test = []
+
+[dependencies]
+serde = "1.0"
+serde_json = { version = "1.0", features = ["raw_value"] }
