--- remote-config/src/dynamic_configuration/mod.rs
@@ -1,4 +1,0 @@
-// Copyright 2021-Present Datadog, Inc. https://www.datadoghq.com/
-// SPDX-License-Identifier: Apache-2.0
-
-pub mod data;
