+++ tools/docker/Dockerfile.build
@@ -82,6 +82,7 @@ COPY "ddcommon-ffi/Cargo.toml" "ddcommon-ffi/"
 COPY "ddtelemetry/Cargo.toml" "ddtelemetry/"
 COPY "ddtelemetry-ffi/Cargo.toml" "ddtelemetry-ffi/"
 COPY "ddsketch/Cargo.toml" "ddsketch/"
+COPY "dynamic-configuration/Cargo.toml" "dynamic-configuration/"
 COPY "profiling/Cargo.toml" "profiling/"
 COPY "profiling-ffi/Cargo.toml" "profiling-ffi/"
 COPY "profiling-replayer/Cargo.toml" "profiling-replayer/"
